Replace height with min-height (trypear/pearai-app#54)
If the pane is too narrow, the content on e.g. Onboarding page sticks over down
from the container because the container's height is fixed.

Just replace `height` with `min-height` and it should fix this.

This was a commit in the archived pearai-submodule. Diffs are removed as part of fixing an incorrect fork